:memo: refreshToken 검사 추가

+++ src/main/java/com/ezel/voza/domain/auth/service/impl/UserLogoutServiceImpl.java
@@ -5,6 +5,7 @@ import com.ezel.voza.domain.auth.repository.RefreshTokenRepository;
 import com.ezel.voza.domain.auth.service.UserLogoutService;
 import com.ezel.voza.global.annotation.ServiceWithTransactional;
 import com.ezel.voza.global.redis.util.RedisUtil;
+import com.ezel.voza.global.security.exception.TokenNotVaildException;
 import com.ezel.voza.global.security.jwt.TokenProvider;
 import lombok.RequiredArgsConstructor;
 
@@ -26,6 +27,10 @@ public class UserLogoutServiceImpl implements UserLogoutService {
         String refresh = tokenProvider.parseToken(refreshToken);
 
         RefreshToken userRefreshToken = refreshTokenRepository.findByRefreshToken(refresh);
+
+        if (userRefreshToken == null) {
+            throw new TokenNotVaildException();
+        }
 
         refreshTokenRepository.delete(userRefreshToken);
